Fixed Paging

+++ sys/kmalloc.c
@@ -12,7 +12,7 @@ void * kmalloc(uint64_t size) {
     if(size > 0x1000) {
          kprintf("KERNEL PANIC: Kmalloc requested page size too big !!!");
     }
-    return (void *)(KERN_BASE + allocate_phys_page());
+    return (void *)((uint64_t)KERN_BASE + allocate_phys_page());
 }
